test(ci): isolate nodes rewards test run to prevent from unrelated rewards
