+++ .github/workflows/nightly.yml
@@ -196,6 +196,12 @@ jobs:
         env:
           SN_LOG: "all"
         timeout-minutes: 10
+
+      - name: execute the nodes rewards tests
+        run: cargo test --release -p sn_node --features="local-discovery" --test nodes_rewards -- --nocapture
+        env:
+          SN_LOG: "all"
+        timeout-minutes: 25
       
       - name: Stop the local network and upload logs
         if: always()
